refactor: junk code and files removed
