+++ WeatherApp/MVC/Model/Locations+CoreDataClass.swift
@@ -1,0 +1,16 @@
+//
+//  Locations+CoreDataClass.swift
+//  
+//
+//  Created by FARIT GATIATULLIN on 15.06.2021.
+//
+//  This file was automatically generated and should not be edited.
+//
+
+//import Foundation
+//import CoreData
+//
+//@objc(Locations)
+//public class Locations: NSManagedObject {
+//
+//}
