+++ WeatherApp/MVC/Model/WeatherApp.xcdatamodeld/Weather.xcdatamodel/contents
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="18154" systemVersion="20F71" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
-    <entity name="Locations" representedClassName="Locations" syncable="YES" codeGenerationType="class">
+    <entity name="Locations" representedClassName="Cities" syncable="YES">
         <attribute name="weatherModel" optional="YES" attributeType="Transformable" customClassName="Locations"/>
     </entity>
     <elements>
