+++ WeatherApp/SwiftUI View/SwiftUIView.swift
@@ -57,7 +57,6 @@ struct ContentView: View {
                 }
             }
         }
-        
     }
 }
