--- WeatherApp/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  WeatherApp
-//
-//  Created by FARIT GATIATULLIN on 08.06.2021.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
